+++ src/main/java/kr/co/morandi/backend/domain/contenttype/random/randomdefense/RandomDefense.java
@@ -1,23 +1,24 @@
-package kr.co.morandi.backend.domain.contenttype.randomdefense;
+package kr.co.morandi.backend.domain.contenttype.random.randomdefense;
 
 import jakarta.persistence.*;
 import kr.co.morandi.backend.domain.contenttype.ContentType;
-import kr.co.morandi.backend.domain.contenttype.randomdefense.randomcriteria.RandomCriteria;
+import kr.co.morandi.backend.domain.contenttype.random.randomcriteria.RandomCriteria;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
 @Entity
-@DiscriminatorValue("StageDefense")
+@DiscriminatorValue("RandomDefense")
 @Getter
 @SuperBuilder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RandomStageDefense extends ContentType {
+public class RandomDefense extends ContentType {
 
     @Embedded
     private RandomCriteria randomCriteria;
 
-    private Double averageStage;
+    private Long problemCount;
 
     private Long timeLimit;
+
 
 }
