+++ src/main/java/kr/co/morandi/backend/domain/contenttype/random/randomdefense/RandomDefenseRepository.java
@@ -1,4 +1,4 @@
-package kr.co.morandi.backend.domain.contenttype.randomdefense;
+package kr.co.morandi.backend.domain.contenttype.random.randomdefense;
 
 import org.springframework.data.jpa.repository.JpaRepository;
