+++ src/main/java/kr/co/morandi/backend/domain/contenttype/random/randomstagedefense/RandomStageDefense.java
@@ -1,24 +1,23 @@
-package kr.co.morandi.backend.domain.contenttype.randomdefense;
+package kr.co.morandi.backend.domain.contenttype.random.randomstagedefense;
 
 import jakarta.persistence.*;
 import kr.co.morandi.backend.domain.contenttype.ContentType;
-import kr.co.morandi.backend.domain.contenttype.randomdefense.randomcriteria.RandomCriteria;
+import kr.co.morandi.backend.domain.contenttype.random.randomcriteria.RandomCriteria;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
 @Entity
-@DiscriminatorValue("RandomDefense")
+@DiscriminatorValue("StageDefense")
 @Getter
 @SuperBuilder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RandomDefense extends ContentType {
+public class RandomStageDefense extends ContentType {
 
     @Embedded
     private RandomCriteria randomCriteria;
 
-    private Long problemCount;
+    private Double averageStage;
 
     private Long timeLimit;
-
 
 }
